refactor(alert): move file in diff folder
