+++ targets/alert-cli/src/__test__/compareTree.test.js
@@ -1,4 +1,4 @@
-import { uniq } from "../compareTree";
+import { uniq } from "../diff/compareDilaTree";
 
 const data = [
   {
